+++ NEWS.md
@@ -1,5 +1,6 @@
 # ellmer (development version)
 
+* `chat_huggingface()` now works much better.
 * [BREAKING CHANGE] `type_array()` and `type_enum()` now have the description as the second argument and `items/`/`values` as the first. This makes them easier to use in the common case where the description isn't necessary (#610).
 * By default, ellmer now retries requests up to 3 times, controllable with `option(ellmer_max_tries)` and will retry if the connection fails (rather than just if the request itself returns a transient error).
 * The default timeout, controlled by `option(ellmer_timeout_s)`, now applies to the initial connection phase.